pyproject.toml
[project]
name = "ots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["starlette[full]>=0.41.0", "uvicorn>=0.32.0"]

[project.optional-dependencies]
dev = ["pytest-asyncio>=0.24.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
